Abstract LinksList component into its own file
