+++ src/components/UI/docs/index.ts
@@ -3,3 +3,4 @@ export * from './Code';
 export * from './DocsLinks';
 export * from './DocsNav';
 export * from './DocumentNav';
+export * from './LinkList';
